lang-trans/requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
deep-translator==1.11.4
langdetect==1.0.9 